Changed input and ouput of date
Added 3 options in yatrc to handle dd/mm/yyyy and mm/dd/yyyy, 12 hour or 24 hour
clock format, and the output format for the date

Changed due_date column from string to float in the database, and the timestamp
is stocked now.

The Due date column has an adaptative length in the ListCommand.