Add a man page, a Changelog and a COPYING file.
In order to ship the man page along with the python packaging, I had to
extend the distutil package. It is pretty painful, given that the doc is
crap and I had to dig into the distutil code in order to get things
working. But now we have a pretty solid system to add some new
documentation.