Add a new mechanism to parse the command line options
The mechanism, for now disabled, can parse options in the Unix style,
short and long. To declare a new option, you just have to add a tuple to
the class variable 'options' of your Command subclass. This tuple
specifies both the short and long versions of the option, the name of
the correspondant object attribute, and a function to translate the one
to the other.

To make the switch, just replace s/cmd.execute/cmd/ in cli/yat