Fix potential bug of unbound method in tree module.
In order to add dynamically a method to a class, you have to set some
internal attributes of the function right, otherwise it might not work
properly, or even fail (most likely scenarii). Or you can use the "new"
module to do it for you.

WARNING : code not tested, non-trivial modification.