Instantiate the yatcli module variables directly in yatcli.
They were previously instantiated in yat, and therefore the statement
"from yatcli import commands" was'nt a sure thing.

Since the str objects are immutable, the only way to change yatcli.name
was to switch instances, which is problematic. To solve this, I renamed
yatcli.name to yatcli.__name and added two functions yatcli.name() and
yatcli.set_name(args) — the args argument was to simplify things ; it
might change in the future.